Move ColorLoader out of recording dir
